refactor app menu
