+++ app/renderer/components/Toolbar/Category/help.ts
