+++ app/renderer/components/Toolbar/store.ts
